Add Support for SDCARD image compression
Compression method to apply to SDCARD after it has been created. 
Supported compression formats are "gzip", "bzip2" or "xz". 
The original .sdcard file is kept and a new compressed file is created if one of these compression formats is chosen. 
If SDCARD_COMPRESSION is set to any other value it is silently ignored.
SDCARD_COMPRESSION variable could be added  in local.conf